[Organization] Move EuiSelectableTemplateSitewide to its own folder
- It was confusing having it amongst the EuiSelectable docs when it's its own completely separate page/documentation

- gave it its own subdir with the same name that it lives in in src/components

- renamed its file names to be clearer and less generic

+ fix onWindowKeyDown  useEffect cleanup removing the wrong event

+ optimize window keyup/keydown fns to not re-init on every rerender
