+++ src-docs/src/views/_index.scss
@@ -11,7 +11,7 @@ $guideDemoHighlightColor: transparentize($euiColorPrimary, .9);
 @import './horizontal_rule/horizontal_rule';
 @import './page/page';
 @import './notification_event/notification_event';
-@import './selectable/search';
+@import './selectable/selectable_templates/sitewide';
 @import './spacer/spacer';
 @import './suggest/index';
 @import './text/text_scaling';
